Rename file and tweaks
